Some cleanups, mostly x86_64 compat plus interlocked macros

+++ pthread_exit.c
@@ -89,7 +89,7 @@ pthread_exit (void *value_ptr)
        */
 
 #if ! defined (__MINGW32__) || defined (__MSVCRT__)  || defined (__DMC__)
-      _endthreadex ((unsigned) value_ptr);
+      _endthreadex ((unsigned) (size_t) value_ptr);
 #else
       _endthread ();
 #endif
